test: 예약 조회 테스트 WebMvcTest로 수정
